refactor class packaging and name of class add documentation.
