+++ src/main/java/com/ideas/rnd/pdf/algo/PdfReportGenerator.java
@@ -1,4 +1,4 @@
-package com.ideas.rnd.pdf;
+package com.ideas.rnd.pdf.algo;
 
 import com.google.common.collect.ObjectArrays;
 import com.ideas.rnd.pdf.model.*;
@@ -17,21 +17,60 @@ import java.io.IOException;
 import java.util.List;
 import java.util.*;
 
-class PDFTableGenerator {
+/**
+ * @author Yogesh Bombe
+ */
+public class PdfReportGenerator {
 	private Table table;
 	private Header header;
 	private Graph graph;
 	private Footer footer;
 	private float tableHeight;
 
-	// Generates document from Table object
-	void generatePDF(Header header, Table table, Graph graph, Footer footer) throws IOException {
+	/**
+	 * Used for restrict to create object using No arg Constructor.
+	 */
+	private PdfReportGenerator() {
+	}
+
+	/**
+	 * @param header
+	 * @param footer
+	 * @param table
+	 */
+	public PdfReportGenerator(Header header, Footer footer, Table table) {
+		this.table = table;
+		this.header = header;
+		this.footer = footer;
+		this.tableHeight = table.isLandscape() ? table.getPageSize().getWidth() - (table.getMargin()) - getHeaderHeight() : table.getPageSize().getHeight() - (table.getMargin()) - getHeaderHeight();
+
+	}
+
+	/**
+	 * @param header
+	 * @param footer
+	 * @param table
+	 * @param graph
+	 */
+	public PdfReportGenerator(Header header, Footer footer, Table table, Graph graph) {
+		assert header != null;
+		this.header = header;
+		assert footer != null;
+		this.footer = footer;
+		assert table != null;
+		this.table = table;
+		assert graph != null;
+		this.graph = graph;
+		this.tableHeight = table.isLandscape() ? table.getPageSize().getWidth() - (table.getMargin()) - getHeaderHeight() : table.getPageSize().getHeight() - (table.getMargin()) - getHeaderHeight();
+	}
+
+	/**
+	 * Generates document from initialize data objects.
+	 *
+	 * @throws IOException
+	 */
+	public void generatePDF() throws IOException {
 		try (PDDocument doc = new PDDocument()) {
-			this.header = header;
-			this.table = table;
-			this.footer = footer;
-			this.graph = graph;
-			this.tableHeight = table.isLandscape() ? table.getPageSize().getWidth() - (table.getMargin()) - getHeaderHeight() : table.getPageSize().getHeight() - (table.getMargin()) - getHeaderHeight();
 			drawTable(doc);
 			if (this.graph != null && this.graph.getGraphs() != null && this.graph.getGraphs().size() > 0) {
 				addGraphs(doc);
@@ -41,20 +80,37 @@ class PDFTableGenerator {
 		}
 	}
 
+	/**
+	 * @return calculated height required for header section.
+	 */
 	private float getHeaderHeight() {
 		return (float) (Math.ceil(this.header.getMetaKeyVal().size() / 2.00f) + 2) * this.table.getRowHeight();
 	}
 
+	/**
+	 * Used to draw graphs
+	 *
+	 * @param doc
+	 * @throws IOException
+	 */
 	private void addGraphs(PDDocument doc) throws IOException {
 		float totalHeightForGraph = this.tableHeight;
 		float totalWidthForGraph = table.getPageSize().getWidth() - (2 * table.getMargin());
 		float tableTopY = this.tableHeight;
 		for (File graphFile : this.graph.getGraphs()) {
-			placeGraph(doc, totalHeightForGraph, totalWidthForGraph, tableTopY, graphFile);
+			drawGraph(doc, totalHeightForGraph, totalWidthForGraph, tableTopY, graphFile);
 		}
 	}
 
-	private void placeGraph(PDDocument doc, float totalHeightForGraph, float totalWidthForGraph, float tableTopY, File graphFile) throws IOException {
+	/**
+	 * @param doc
+	 * @param totalHeightForGraph
+	 * @param totalWidthForGraph
+	 * @param tableTopY
+	 * @param graphFile
+	 * @throws IOException
+	 */
+	private void drawGraph(PDDocument doc, float totalHeightForGraph, float totalWidthForGraph, float tableTopY, File graphFile) throws IOException {
 		PDPage page = generatePage(doc);
 		PDPageContentStream contentStream = generateContentStream(doc, page);
 		addHeader(contentStream);
@@ -66,6 +122,11 @@ class PDFTableGenerator {
 		contentStream.close();
 	}
 
+	/**
+	 * @param total
+	 * @param offset
+	 * @return adjusted height and width of graph.
+	 */
 	private float getGraphAdjustmentScale(float total, int offset) {
 		float width;
 		if (offset > total) {
@@ -76,7 +137,12 @@ class PDFTableGenerator {
 		return width;
 	}
 
-	private float addHeader(PDPageContentStream contentStream) throws IOException {
+	/**
+	 * @param contentStream
+	 * @return
+	 * @throws IOException
+	 */
+	private void addHeader(PDPageContentStream contentStream) throws IOException {
 		float xPos = table.getMargin();
 		float totalWidth = table.getPageSize().getWidth() - (table.getMargin() * 2);
 		float cellWidth = totalWidth / 4;
@@ -120,9 +186,12 @@ class PDFTableGenerator {
 			}
 			flag++;
 		}
-		return nextY;
 	}
 
+	/**
+	 * @param doc
+	 * @throws IOException
+	 */
 	private void addFooter(PDDocument doc) throws IOException {
 		int numberOfPages = doc.getNumberOfPages();
 		for (int pageNumber = 0; pageNumber < numberOfPages; pageNumber++) {
@@ -135,6 +204,11 @@ class PDFTableGenerator {
 		}
 	}
 
+	/**
+	 * @param doc
+	 * @param footerContentStream
+	 * @throws IOException
+	 */
 	private void drawHorizontalLine(PDDocument doc, PDPageContentStream footerContentStream) throws IOException {
 		footerContentStream.setStrokingColor(this.footer.getLineColor() != null ? this.footer.getLineColor() : Color.LIGHT_GRAY);
 		footerContentStream.setLineWidth(this.footer.getLineWidth());
@@ -148,21 +222,37 @@ class PDFTableGenerator {
 		footerContentStream.setStrokingColor(Color.BLACK);
 	}
 
+	/**
+	 * @param doc
+	 * @param footerContentStream
+	 * @throws IOException
+	 */
 	private void drawLeftSection(PDDocument doc, PDPageContentStream footerContentStream) throws IOException {
 		PDImageXObject pdImage = PDImageXObject.createFromFileByExtension(this.footer.getLogoImage(), doc);
 		footerContentStream.drawImage(pdImage, table.getPageSize().getLowerLeftX() + table.getMargin(), table.getPageSize().getLowerLeftY() + 20, 30, 15);
 	}
 
+	/**
+	 * @param numberOfPages
+	 * @param pageNumber
+	 * @param footerContentStream
+	 * @throws IOException
+	 */
 	private void drawRightSection(int numberOfPages, int pageNumber, PDPageContentStream footerContentStream) throws IOException {
 		footerContentStream.beginText();
 		footerContentStream.setFont(PDType1Font.TIMES_ROMAN, 8);
 		footerContentStream.newLineAtOffset(table.getPageSize().getUpperRightX() - table.getMargin() - 40, table.getPageSize().getLowerLeftY() + 20);
 		footerContentStream.showText(String.format(this.footer.getPageNumberPhrase(), pageNumber, numberOfPages));
-		//footerContentStream.showText("Page " + pageNumber + " of " + numberOfPages);
 		footerContentStream.endText();
 	}
 
-	// Configures basic setup for the table and draws it page by page
+
+	/**
+	 * Configures basic setup for the table and draws it page by page
+	 *
+	 * @param doc
+	 * @throws IOException
+	 */
 	private void drawTable(PDDocument doc) throws IOException {
 		// Calculate pagination
 		Integer rowsPerPage;
@@ -192,6 +282,9 @@ class PDFTableGenerator {
 		}
 	}
 
+	/**
+	 * @return calculated table column configuration for column split.
+	 */
 	private List<List<Range>> getSinglePageRange() {
 		List<List<Range>> lists = new ArrayList<>();
 		List<Range> ranges = new ArrayList<>();
@@ -206,6 +299,10 @@ class PDFTableGenerator {
 		return lists;
 	}
 
+	/**
+	 * @param isFixedColumn
+	 * @return
+	 */
 	private List<List<Range>> getRangesOfColumnRangePerPageNew(boolean isFixedColumn) {
 		List<List<Range>> listList = new ArrayList<>();
 		float totalWidth = table.getPageSize().getWidth() - (table.getMargin() * 2);
@@ -258,6 +355,13 @@ class PDFTableGenerator {
 
 	}
 
+	/**
+	 * @param listList
+	 * @param fixedColumns
+	 * @param isFixedColumn
+	 * @param numberOfColumns
+	 * @param lastOffset
+	 */
 	private void getLastOffsetNew(List<List<Range>> listList, List<Range> fixedColumns, boolean isFixedColumn, Integer numberOfColumns, float lastOffset) {
 		List<Range> ranges1 = listList.get(listList.size() - 1);
 		int to = ranges1.get(ranges1.size() - 1).getTo();
@@ -271,6 +375,12 @@ class PDFTableGenerator {
 		listList.add(ranges);
 	}
 
+	/**
+	 * @param numberOfColumns
+	 * @param lastOffset
+	 * @param i
+	 * @return
+	 */
 	private Range getRange(Integer numberOfColumns, float lastOffset, int i) {
 		Range range = new Range();
 		range.setFrom(i);
@@ -279,6 +389,12 @@ class PDFTableGenerator {
 		return range;
 	}
 
+	/**
+	 * @param columns
+	 * @param from
+	 * @param to
+	 * @return
+	 */
 	private float getOffset(List<Column> columns, int from, int to) {
 		float fixedColumnWidth = 0;
 		for (int i = from; i < to + 1; i++) {
@@ -287,11 +403,16 @@ class PDFTableGenerator {
 		return fixedColumnWidth;
 	}
 
+	/**
+	 * @return number of rows per page.
+	 */
 	private Integer getRowsPerPage() {
 		return new Double(Math.floor((this.tableHeight - this.table.getMargin() - table.getColumnHeight()) / table.getRowHeight())).intValue();
 	}
 
-	// Draws current page table grid and border lines and content
+	/**
+	 * Draws current page table grid and border lines and content
+	 */
 	private void drawCurrentPage(List<List<String>> currentPageContent, PDPageContentStream contentStream, List<Range> ranges, boolean isFixedColumn)
 			throws IOException {
 		float nextX = 0;
@@ -335,7 +456,19 @@ class PDFTableGenerator {
 		contentStream.close();
 	}
 
-	// Writes the content for one line
+	/**
+	 * Writes the content for one line
+	 *
+	 * @param lineContent
+	 * @param contentStream
+	 * @param nextTextX
+	 * @param nextTextY
+	 * @param table
+	 * @param range
+	 * @param to
+	 * @param isFixedColumn
+	 * @throws IOException
+	 */
 	private void writeContentLine(String[] lineContent, PDPageContentStream contentStream, float nextTextX, float nextTextY,
 								  Table table, Range range, int to, boolean isFixedColumn) throws IOException {
 		int from = range.getFrom();
@@ -366,7 +499,9 @@ class PDFTableGenerator {
 	}
 
 
-	// Writes the content for one line
+	/**
+	 * Writes the content for one line
+	 */
 	private void writeHeaderContentLine(String[] lineContent, PDPageContentStream contentStream, float nextTextX, float nextTextY,
 										Range range) throws IOException {
 		int from = range.getFrom();
@@ -390,6 +525,17 @@ class PDFTableGenerator {
 	}
 
 
+	/**
+	 * @param table
+	 * @param currentPageContent
+	 * @param contentStream
+	 * @param tableTopY
+	 * @param range
+	 * @param expNextX
+	 * @param isFixedColumn
+	 * @return
+	 * @throws IOException
+	 */
 	private float drawTableGrid(Table table, List<List<String>> currentPageContent, PDPageContentStream contentStream, float tableTopY, Range range, float expNextX, boolean isFixedColumn)
 			throws IOException {
 		float nextY = tableTopY;
@@ -429,6 +575,13 @@ class PDFTableGenerator {
 		return nextX;
 	}
 
+	/**
+	 * @param range
+	 * @param expNextX
+	 * @param isFixedColumn
+	 * @param nextX
+	 * @return
+	 */
 	private float getNextX(Range range, float expNextX, boolean isFixedColumn, float nextX) {
 		if (range.getFrom() != 0 && isFixedColumn) {
 			nextX = expNextX;
@@ -437,6 +590,12 @@ class PDFTableGenerator {
 	}
 
 
+	/**
+	 * @param rowsPerPage
+	 * @param pageCount
+	 * @param ranges
+	 * @return
+	 */
 	private List<List<String>> getContentForCurrentPage(Integer rowsPerPage, int pageCount, List<Range> ranges) {
 		List<List<String>> newContentList = new ArrayList<>();
 
@@ -458,6 +617,10 @@ class PDFTableGenerator {
 		return newContentList;
 	}
 
+	/**
+	 * @param doc
+	 * @return
+	 */
 	private PDPage generatePage(PDDocument doc) {
 		PDPage page = new PDPage();
 		page.setMediaBox(table.getPageSize());
@@ -466,17 +629,32 @@ class PDFTableGenerator {
 		return page;
 	}
 
+	/**
+	 * @param doc
+	 * @param page
+	 * @return
+	 * @throws IOException
+	 */
 	private PDPageContentStream generateContentStream(PDDocument doc, PDPage page) throws IOException {
 		PDPageContentStream contentStream = new PDPageContentStream(doc, page, PDPageContentStream.AppendMode.PREPEND, true, true);
 		if (table.isLandscape()) {
 			Matrix matrix = new Matrix(0, 1, -1, 0, table.getPageSize().getWidth(), 0);
 			contentStream.transform(matrix);
 		}
-		//contentStream.setFont(table.getContentTextFont(), table.getContentFontSize());
 		return contentStream;
 	}
 
 
+	/**
+	 * @param contentStream
+	 * @param lineColor
+	 * @param lineWidth
+	 * @param startX
+	 * @param endX
+	 * @param startY
+	 * @param endY
+	 * @throws IOException
+	 */
 	private void drawLine(PDPageContentStream contentStream, Color lineColor, float lineWidth, float startX, float endX, float startY, float endY) throws IOException {
 		contentStream.setStrokingColor(lineColor);
 		contentStream.setLineWidth(lineWidth);
@@ -487,6 +665,15 @@ class PDFTableGenerator {
 		contentStream.setStrokingColor(Color.BLACK);
 	}
 
+	/**
+	 * @param contentStream
+	 * @param startX
+	 * @param startY
+	 * @param width
+	 * @param height
+	 * @param color
+	 * @throws IOException
+	 */
 	private void drawCellBackground(PDPageContentStream contentStream, final float startX, final float startY, final float width, final float height, Color color)
 			throws IOException {
 		contentStream.setNonStrokingColor(color);
@@ -497,6 +684,16 @@ class PDFTableGenerator {
 		contentStream.setNonStrokingColor(Color.BLACK);
 	}
 
+	/**
+	 * @param contentStream
+	 * @param textColor
+	 * @param textFont
+	 * @param fontSize
+	 * @param y
+	 * @param x
+	 * @param text
+	 * @throws IOException
+	 */
 	private void writeText(PDPageContentStream contentStream, Color textColor, PDFont textFont, float fontSize, float y, float x, String text) throws IOException {
 		contentStream.beginText();
 		contentStream.setNonStrokingColor(textColor);
@@ -506,6 +703,13 @@ class PDFTableGenerator {
 		contentStream.endText();
 	}
 
+	/**
+	 * @param textAlignment
+	 * @param cellWidth
+	 * @param cellMargin
+	 * @param textWidth
+	 * @return
+	 */
 	private float getAdjustX(Alignment textAlignment, float cellWidth, float cellMargin, int textWidth) {
 		Alignment alignment = textAlignment == null ?
 				Alignment.LEFT : textAlignment;
@@ -527,14 +731,33 @@ class PDFTableGenerator {
 		return adjustment;
 	}
 
+	/**
+	 * @param cellWidth
+	 * @param cellMargin
+	 * @param textWidth
+	 * @return
+	 */
 	private float getCenterX(float cellWidth, float cellMargin, int textWidth) {
 		return (cellWidth - cellMargin - textWidth) / 2;
 	}
 
+	/**
+	 * @param cellWidth
+	 * @param cellMargin
+	 * @param textWidth
+	 * @return
+	 */
 	private float getRightX(float cellWidth, float cellMargin, int textWidth) {
 		return (cellWidth - (textWidth + cellMargin * 2));
 	}
 
+	/**
+	 * @param textFont
+	 * @param text
+	 * @param fontSize
+	 * @return
+	 * @throws IOException
+	 */
 	private int getTextWidth(PDFont textFont, String text, float fontSize) throws IOException {
 		return (int) ((textFont.getStringWidth(text) / 1000) * fontSize);
 	}
